feat!: rename imageProcessing middleware and skip processing for non-image types
REHLTYEXT-85
